moving facts about drugs block to stand alone component to avoid PageComponent re-render
