+++ app/shared/components/MatomoAnalytics/component.jsx
@@ -1,15 +1,12 @@
 import React from 'react'
-import MatomoTracker from '@datapunt/matomo-tracker-js'
 import { getMatomoTracker } from '../../utilities'
 
 const MatomoAnalytics = () => {
   if (typeof window !== 'undefined') {
     const tracker = getMatomoTracker()
-    console.log('tracking page')
     tracker.trackPageView()
   }
-  return (
-    null
-  )
+  return null
 }
+
 export default MatomoAnalytics
